import-data/requirements.txt
chardet==5.2.0
dotenv==0.9.9
numpy==2.3.1
pandas==2.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
